+++ src/utils/date/tests/date.test.js
@@ -1,30 +1,11 @@
 import * as date from '../date';
 
-const appendZero = (val) => (val > 9 ? val : `0${val}`);
-const createTestResult = (val) => {
-  const dateObj = val instanceof Date ? val : new Date(val);
-
-  const year = dateObj.getFullYear();
-  const month = dateObj.getMonth() + 1;
-  const adjustedMonth = appendZero(month);
-  const dates = dateObj.getDate();
-  const adjustedDate = appendZero(dates);
-  const hour = dateObj.getHours();
-  const adjustedHour = appendZero(hour);
-  const minute = dateObj.getMinutes();
-  const adjustedMinute = appendZero(minute);
-  const second = dateObj.getSeconds();
-  const adjustedSecond = appendZero(second);
-
-  return `${year}-${adjustedMonth}-${adjustedDate} ${adjustedHour}:${adjustedMinute}:${adjustedSecond}`;
-};
-
 describe('date helper', () => {
   describe('formatDateFromObject', () => {
     it('construct correct date string if date object is passed', () => {
       const correctInitialParams = new Date('Sat May 02 2020 14:25:05 GMT+0200');
 
-      const expectedResult = createTestResult(correctInitialParams);
+      const expectedResult = date.createDateForTestResult(correctInitialParams);
 
       const result = date.formatDateFromObject(correctInitialParams);
 
@@ -35,7 +16,7 @@ describe('date helper', () => {
     it('construct correct date string if date string is passed', () => {
       const correctInitialParams = '2020-01-20T15:01:58.957Z';
 
-      const expectedResult = createTestResult(correctInitialParams);
+      const expectedResult = date.createDateForTestResult(correctInitialParams);
 
       const result = date.formatDateFromString(correctInitialParams);
